% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create the result table for glmnet models}
\usage{
create_table(model0, models, nams, errors0, errors, debiased_errors0,
  debiased_errors, index, S, p, pvalue, selection_frequency, model_score)
}
\arguments{
\item{model0}{models for the original procedure}

\item{models}{models for the proximal procedures}

\item{nams}{a length p vector containing feature names.}

\item{errors0}{the n by m errors of the original model sequence}

\item{errors}{a list of n by m errors of the proximal model sequences}

\item{debiased_errors0}{debiased cross-validation/test error estimate for selected model in the original model sequence}

\item{debiased_errors}{debiased cross-validation/test error estimates for selected model in the alternative model sequences}

\item{index}{penalty index for those representative models}

\item{S}{selected features}

\item{p}{size of full feature dimension}

\item{pvalue}{model pvalues, default is NULL}

\item{selection_frequency}{selection frequencies, default is NULL}

\item{model_score}{model scores, default is NULL}
}
\value{
resultTable: organized result table for proximity analysis
}
\description{
Create the result table for glmnet models
}
