% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_bootstrap.R
\name{fastBoostrap}
\alias{fastBoostrap}
\title{Default Bootstrap}
\usage{
fastBoostrap(n, B)
}
\arguments{
\item{n}{number of samples}

\item{B}{number of Bootstrap repetitions}
}
\value{
A: 0 indexes n by B Bootstrap index matrix
}
\description{
Default Bootstrap
}
