% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndex}
\alias{getIndex}
\title{Perform the model index selection}
\usage{
getIndex(errors0, cv_glm, alpha = 0.1, epsilon = 0.05^2,
  selectionType = 0, one_sds = rep(0, ncol(errors0)))
}
\arguments{
\item{errors0}{the n by m errors of the original model sequence}

\item{cv_glm}{Return from the function cv.glmnet keep = T.}

\item{alpha}{added error with covariance structure*alpha, by default, alpha = .1}

\item{epsilon}{added error with covariance structure being identity times min(covariance diagonal) times epsilon, by default, epsilon = 0.05^2}

\item{selectionType}{if selectionType == 0, pick the model with the smallest randomized error}

\item{one_sds}{if the selectionType is 1, the we choose the model with smallest index
such that model error(randomized) <= one_sds[i] + min error(randomized)}
}
\value{
model_index: de-biased estimate of the model error for the original procedure
}
\description{
This function can be used to evaluate selection frequency for general training procedure with errors0 be the error matrix using original procedure with the Bootstrap samples with the same sequence of penalties
}
