% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextdoor_test.R
\name{nextdoor_unconditional_test}
\alias{nextdoor_unconditional_test}
\title{Provide Bootstrap p-values}
\usage{
nextdoor_unconditional_test(errors0, errors, S, nams = NULL,
  random_estimates, sigmaLeft, epsilon = NULL, epsilon2 = NULL,
  alpha = 0.1, K = 100, B = 1000, Bindex = NULL, rescale = T,
  selectionType = 0, one_sds = rep(0, ncol(errors0)), trace = T)
}
\arguments{
\item{errors0}{the n by m errors of the original model sequence}

\item{errors}{a list of n by m errors of the proximal model sequences}

\item{S}{a vector corresponding of feature indexes considered in the nextdoor models.}

\item{nams}{a length p vector containing feature names.}

\item{random_estimates}{the randomized error estimate of the deterioration proximal - original}

\item{sigmaLeft}{Left eigenvector matrix for errors of each of the model sequence.}

\item{epsilon}{added error with covariance structure being identity times min(covariance diagonal) times epsilon,
by default, epsilon = 0}

\item{epsilon2}{added error in the Bootstrap step being min(covariance diagonal) times epsilon2,
by default, epsilon2 = 0}

\item{alpha}{added error with covariance structure*alpha, by default, alpha = .1}

\item{K}{number of repetitions estimating the de-biased error}

\item{B}{number of bootstrap repetitions}

\item{Bindex}{n by B index matrix for bootstrap. if Bindex == NULL,
use the default Bootstrap, otherwise, use the provide matrix}

\item{rescale}{if rescale == True, perform the mean-rescaled Bootstrap}

\item{selectionType}{if selectionType == 0, pick the model with the smallest randomized error
if selectionType == 1, use the 1se rule}

\item{one_sds}{if the selectionType is 1, the we choose the model with smallest index
such that model error(randomized) <= one_sds[i] + min error(randomized)}

\item{trace}{if trace == True, print the p-value process}
}
\value{
pvalues: p values for nextdoor analysis
}
\description{
Provide Bootstrap p-values
}
