% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nextdoor}
\alias{print.nextdoor}
\title{Print result summary of a nextdoor object}
\usage{
\method{print}{nextdoor}(x, ...)
}
\arguments{
\item{x}{return from nextdoor.glmnet or nextdoor.}

\item{...}{parameters for print.}
}
\description{
Print result summary of a nextdoor object
}
