% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model2}
\alias{train_model2}
\title{Perform training, model selection}
\usage{
train_model2(x, y, family = c("gaussian", "binomial", "poisson",
  "multinomial"), foldid = NULL, lambda = NULL, lambda_extra = NULL,
  lossfun = NULL, standardize = FALSE, epsilon, alpha, selectionType)
}
\arguments{
\item{x}{n by p training feature matrix}

\item{y}{length Y response vector}

\item{family}{model family}

\item{foldid}{fold id for across-validation, if it is NULL, the samples are randomly divided into nfolds folds.}

\item{lambda}{A user supplied lambda sequence. WARNING: use with care. Avoid supplying a single value for lambda.}

\item{lambda_extra}{lambda padded with a small value.}

\item{lossfun}{The user provided loss function used cross-validation. It has the form lossfun(Y_pred, Y),
with Y_pred being the predicted values using the link function. If lossfun == NULL, by default, we use the deviance.}

\item{standardize}{whether to standardize the data, standardize = T by default}

\item{epsilon}{added error with covariance structure being identity times min(covariance diagonal) times epsilon, by default, epsilon = 0.05^2}

\item{alpha}{added error with covariance structure*alpha, by default, alpha = .1}

\item{selectionType}{if selectionType == 0, pick the model with the smallest randomized error
if selectionType == 1, use the 1se rule}
}
\value{
S: selected features indexes.

index0: selected model index.
}
\description{
Perform training, model selection
}
